mostly done with the style/layout of the website
In this commit: the game is completely and correctly functional to play a endless game of rock paper scissors. The css is at an acceptable commit point, the game is centered correctly with the correct heading below the pictures. The color scheme is not the best, odds are I will change it before I consider this project complete. The javascript does technically work if my goal is to just have it play an endless game of rps, I might work on making it to where it is the first one to 5 and the score is kept.